+++ src/main/java/presenter/TransitMapPresenter.java
@@ -1,5 +1,4 @@
 package presenter;
 
 public interface TransitMapPresenter {
-    public void
 }
