+++ src/main/java/ui/WelcomePage.java
@@ -2,7 +2,7 @@ package ui;
 
 import ui.passenger.PassengerHomePage;
 import ui.round.RoundedButton;
-import ui.staff.StaffHomePage;
+import ui.staff.StaffSelectPage;
 
 import javax.swing.*;
 import java.awt.*;
@@ -57,7 +57,7 @@ public class WelcomePage {
         staffButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                new StaffHomePage();
+                new StaffSelectPage();
                 frame.dispose();
             }
         });
