+++ src/main/java/ui/colorTester.java
@@ -3,11 +3,6 @@ package ui;
 import javax.swing.*;
 
 public class colorTester {
-    private JButton button1;
-    private JButton button2;
-    private JButton button3;
-    private JButton button4;
-    private JButton button5;
-    private JButton button6;
-    private JButton button7;
+    private JPanel panel1;
+    private JScrollBar scrollBar1;
 }
