--- src/main/java/ui/staff/admin/AdminMapPage.java
@@ -1,4 +1,0 @@
-package ui.staff.admin;
-
-public class AdminMapPage {
-}
