+++ src/main/java/usecase/TransitState.java
@@ -1,5 +1,4 @@
 package usecase;
 
 public class TransitState {
-    private
 }
